move dmx code out of break.py
