+++ src/break.py
@@ -2,7 +2,6 @@ import time
 from datetime import datetime
 from multiprocessing import Process
 
-import blink
 import display
 import dtxpro
 import keyboard
@@ -10,20 +9,18 @@ import keys
 import lights
 import sample
 import utility
-from dtxpro import DtxPad
-from sequence import sequence
-
-import midi
 
 # import rtmidi
 # from ctypes import *
 # from contextlib import contextmanager
-
 # import display
 # import lights
-
 # from dmx import DMXInterface
+from blink import update_dmx
+from dtxpro import DtxPad
+from sequence import sequence
 
+import midi
 
 # with DMXInterface() as interface:
 #
@@ -32,12 +29,6 @@ import midi
 logger = utility.get_logger(__name__)
 current_time = datetime.now().strftime("%H:%M:%S")
 logger.debug(f"Start time = {current_time}")
-
-dmx_interface = None
-# try:
-#     dmx_interface = DMXInterface()
-# except:
-#     pass
 
 
 def on_key(e):
@@ -54,74 +45,13 @@ keyboard.hook(on_key)
 sample.init()
 display.init(sample.all_samples())
 midi.connect()
-midi.load_midi_files()
+# midi.load_midi_files()
 # sequence.control_bpm(control.encoder)
-
-
-def bounce(step):
-    x = step % 32
-    if x < 16:  # noqa: PLR2004
-        return x + 1
-    return 16 - (x % 16)
-
-
-def bounce_lights(step):
-    # tri = modulation.triangle(15)
-    # light = round(tri(step % 16) * 7)
-    light = bounce(step)
-    logger.debug(f"light {light}")
-    return (light,)
-
-
-def lights_for_step(step):
-    light_index = step % 8 + 1 + 8
-    mirror_index = -(light_index - 8) + 8 + 1
-    return (light_index, mirror_index)
 
 # if (now := time.time()) - last_dmx > 0.050: # and last_dmx_step != sequence.step:
 # TODO move into own file
-
-
-def update_dmx(step, note_number=None):
-    if dmx_interface is None:
-        return
-    # last_dmx = now
-    # last_dmx_step = sequence.step
-    logger.debug(f"lighting dmx step {step}")
-    color = [0, 0, 0, 0, 0, 0]
-    time.sleep(0.020)
-
-    blink.Light.scale(0.8)
-
-    for i, s in enumerate(sample.current_samples()):
-        if s.channel and s.channel.get_busy():
-            source_step = sample.sound_data[s.channel.get_sound()].source_step
-            if source_step != sequence.step:
-                logger.debug(f"source step {source_step}")
-            color[i] = 255
-            for j in bounce_lights(source_step):
-                blink.lights[(j + i * 3) % len(blink.lights)].absorb(color)
-        # blink.lights[i].absorb([0,0,0,0,0,0])
-
-    for note_number in midi.note_q:
-        if note_number == 0:
-            for light in blink.lights:
-                # logger.info(f"bass flash {step} {note_number}")
-                light.absorb([255, 0, 0, 255, 255, 255])
-    midi.note_q.clear()
-
-    # light_index = sequence.step % 8 + 1 + 8
-    # mirror_index = -(light_index - 8) + 8 + 1
-    # if any(color):
-    #     blink.lights[light_index].set(color)
-    #     blink.lights[mirror_index].set(color)
-    dmx_interface.set_frame(list(blink.Light.data))
-    time.time()
-    dmx_interface.send_update()
-
-
-    # sample.Sample.audio_executor.submit(dmx_interface.send_update)
-    # logger.info(f"dmx frame send took {time.time() - now}s")
+# sample.Sample.audio_executor.submit(dmx_interface.send_update)
+# logger.info(f"dmx frame send took {time.time() - now}s")
 sequence.on_step(lambda s: sample.Sample.audio_executor.submit(update_dmx, s))
 
 subs = [
